feat: [backend-lint-size-003] - [Bring maintained backend test code under the size limits]
